+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-08-16 09:54:05
-// Commit: chore(security): update dependencies and cleanup unused imports in security
+// Updated at: 2026-08-16 10:54:31
+// Commit: feat(security): add bi-directional event bus handler for security
 
 class securityComponentState {
-  final String commitRef = '3623';
-  final String lastMessage = 'chore(security): update dependencies and cleanup unused imports in security';
-  final DateTime timestamp = DateTime.parse('2026-08-16T09:54:05+05:00');
+  final String commitRef = '3630';
+  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
+  final DateTime timestamp = DateTime.parse('2026-08-16T10:54:31+05:00');
 }